+++ src/extension/app/store/app.js
@@ -455,7 +455,7 @@ export class AppStore {
    */
   showWait(message) {
     if (!message) {
-      message = i18n(this.languageDict, 'please_wait');
+      message = this.i18n('please_wait');
     }
     EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
       detail: {
@@ -470,6 +470,14 @@ export class AppStore {
    */
   hideWait() {
     EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.CLOSE_MODAL));
+  }
+
+  /**
+   * Reloads the current page. Abstracted for testing.
+   */
+  reloadPage() {
+    // istanbul ignore next
+    window.location.reload();
   }
 
   /**
@@ -636,6 +644,15 @@ export class AppStore {
   }
 
   /**
+   * Helper i18n function
+   * @param {string} key Dictionary key
+   * @returns {string} The translated string
+   */
+  i18n(key) {
+    return i18n(this.languageDict, key);
+  }
+
+  /**
    * Updates the observable status of the current resource.
    * @param {Object} status The status object
    */
@@ -706,7 +723,7 @@ export class AppStore {
           detail: {
             type: MODALS.ERROR,
             data: {
-              message: `${i18n(this.languageDict, 'error_config_failure')}${resp.error}`,
+              message: `${this.i18n('error_config_failure')}${resp.error}`,
             },
           },
         }));
@@ -718,7 +735,7 @@ export class AppStore {
           detail: {
             type: MODALS.ERROR,
             data: {
-              message: i18n(this.languageDict, 'error_preview_failure'),
+              message: this.i18n('error_preview_failure'),
             },
           },
         }));
@@ -727,7 +744,7 @@ export class AppStore {
     }
     // handle special case /.helix/*
     if (status.webPath.startsWith('/.helix/')) {
-      this.showToast(i18n(this.languageDict, 'preview_config_success'), 'positive');
+      this.showToast(this.i18n('preview_config_success'), 'positive');
       return;
     }
     this.hideWait();
@@ -767,6 +784,7 @@ export class AppStore {
           ...getAdminFetchOptions(),
         },
       );
+
       // bust client cache for live and production
       if (siteStore.outerHost) {
         // reuse purgeURL to ensure page relative paths (e.g. when publishing dependencies)
@@ -793,7 +811,7 @@ export class AppStore {
    * @param {boolean} [open] true if environment should be opened in new tab
    * @fires Sidekick#envswitched
    */
-  async switchEnv(targetEnv, open = false) {
+  switchEnv(targetEnv, open = false) {
     this.showWait();
     const hostType = ENVS[targetEnv];
     if (!hostType) {