+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+# [1.7.0](https://github.com/adobe/aem-sidekick/compare/v1.6.2...v1.7.0) (2024-2-7)
+
+
+### Bug Fixes
+
+* backward compatibility ([f82adc4](https://github.com/adobe/aem-sidekick/commit/f82adc4c7d6d080c90cab19bdafd2389e5c45b31))
+
+
+### Features
+
+* view doc source ([dc6e545](https://github.com/adobe/aem-sidekick/commit/dc6e54522f2269a8db5d09f6ccfc26b20994e4c8))
+
+## [1.6.2](https://github.com/adobe/aem-sidekick/compare/v1.6.1...v1.6.2) (2024-2-2)
+
+
+### Bug Fixes
+
+* add publish tests ([28a8889](https://github.com/adobe/aem-sidekick/commit/28a8889f139f7e6f5ed8db8292e800c1af252702))
+* attempt to fix restore ([446c986](https://github.com/adobe/aem-sidekick/commit/446c986e5ad45a9e671ce04c0c179e7adab78b62))
+* cleanup mocha tests ([c008c66](https://github.com/adobe/aem-sidekick/commit/c008c667f4d7e16c20c004a17777540ed3bce946))
+* move preview tests out of wtr ([feb37d1](https://github.com/adobe/aem-sidekick/commit/feb37d1164c90796970daafe69bdcb8d51b0e603))
+* move publish tests out of wtr directory ([93379ed](https://github.com/adobe/aem-sidekick/commit/93379ed27892639144cadd8356cb388a4b8dec68))
+* move tests into root test directory ([34670e5](https://github.com/adobe/aem-sidekick/commit/34670e59f1df8aeeb689709f40dc5e05a07340cc))
+* typedef import path ([9d50062](https://github.com/adobe/aem-sidekick/commit/9d500629476a84b726358fc4ce55366e44ef0074))
+
+## [1.6.1](https://github.com/adobe/aem-sidekick/compare/v1.6.0...v1.6.1) (2024-2-2)
+
+
+### Bug Fixes
+
+* abstract window.location.reload for testing ([bfe873b](https://github.com/adobe/aem-sidekick/commit/bfe873b2d088b615c1303ee7f1a853b0429ce00a))
+* call click method instead of event dispatching ([3bdc760](https://github.com/adobe/aem-sidekick/commit/3bdc760a20ad98d31aba30f31801af4c98d69e51))
+* ignore window.location.reload wrapper in appstore ([61d4eb3](https://github.com/adobe/aem-sidekick/commit/61d4eb394a5586db7f6d701b582c98ae41808b42))
+* preview plugin tests ([5eb3599](https://github.com/adobe/aem-sidekick/commit/5eb3599ee0f1f49d74677f2c6380075bc868e67f))
+* support disabled state in core plugins ([7948635](https://github.com/adobe/aem-sidekick/commit/7948635a3a760e6571cc0d637c5763dec45a9c5e))
+* update and updatePreview coverage in appStore ([ed4ae28](https://github.com/adobe/aem-sidekick/commit/ed4ae287e391cfc6d3399e1d6e2b90fa08fda9b7))
+* update test fixtures to include payloads for gdrive and sharepoint ([b51fe2f](https://github.com/adobe/aem-sidekick/commit/b51fe2fd858214d10231cf7d888567f0bdcc437c))
+
+# [1.6.0](https://github.com/adobe/aem-sidekick/compare/v1.5.3...v1.6.0) (2024-1-29)
+
+
+### Features
+
+* tweak env switcher ([aad88ee](https://github.com/adobe/aem-sidekick/commit/aad88eed0ed4c5d36cad2e7025f68c5a0072126c))
+
+## [1.5.3](https://github.com/adobe/aem-sidekick/compare/v1.5.2...v1.5.3) (2024-1-27)
+
+
+### Bug Fixes
+
+* add and use i18n helper function on appStore ([06bb150](https://github.com/adobe/aem-sidekick/commit/06bb150cb7ae2f21758fcbcbfb40e58d2ee572c6))
+
 ## [1.5.2](https://github.com/adobe/aem-sidekick/compare/v1.5.1...v1.5.2) (2024-1-26)